requirements.txt
ansi2html==1.6.0
argon2-cffi @ file:///C:/ci/argon2-cffi_1613037869401/work
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
backcall==0.2.0
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
Bottleneck @ file:///C:/ci/bottleneck_1607557040328/work
Brotli==1.0.9
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831756778/work
charset-normalizer==2.0.5
click==8.0.1
colorama @ file:///tmp/build/80754af9/colorama_1603211150991/work
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy @ file:///C:/ci/debugpy_1629222819322/work
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==2.0.1
Flask-Compress==1.10.1
idna==3.2
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///C:/ci/ipykernel_1629995228567/work/dist/ipykernel-6.2.0-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1628244015142/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
itsdangerous==2.0.1
jedi @ file:///C:/ci/jedi_1611341083684/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1621419064915/work
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter @ file:///C:/ci/jupyter_1607685287094/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213515523/work
jupyter-dash==0.4.0
MarkupSafe @ file:///C:/ci/markupsafe_1621528502553/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune @ file:///C:/ci/mistune_1607359457024/work
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
nbconvert==5.5.0
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
notebook @ file:///C:/ci/notebook_1629205890343/work
numexpr @ file:///C:/ci/numexpr_1618856728739/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1626271900803/work
olefile==0.46
packaging==20.4
pandas @ file:///C:/ci/pandas_1629682571109/work
pandocfilters @ file:///C:/ci/pandocfilters_1605114832805/work
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1625663294916/work
plotly==5.3.1
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyparsing==2.4.7
pyrsistent @ file:///C:/ci/pyrsistent_1625052426389/work
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty @ file:///C:/ci/pywinpty_1607419945780/work
pyzmq @ file:///C:/ci/pyzmq_1628276105793/work
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
requests==2.26.0
retrying==1.3.3
scikit-learn @ file:///D:/bld/scikit-learn_1605985813794/work
scipy @ file:///C:/ci/scipy_1630606917240/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
tenacity==8.0.1
terminado==0.9.4
testpath==0.4.4
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tornado @ file:///C:/ci/tornado_1606924294691/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
urllib3==1.26.6
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension @ file:///C:/ci/widgetsnbextension_1607531582688/work
wincertstore==0.2
zipp @ file:///tmp/build/80754af9/zipp_1603306325174/work
